[CP][EWT-797] Fixing max_active_runs to control the number of active runs
* EWT-797: Fixing max_active_runs to control the number of runs running simultaneously

* Bumped the version

Co-authored-by: Shruti Mantri <smantri@twitter.com>

I referred to the git issue: apache#9975 which in its first line itself tells that the issue that happens related to max_active_runs in 1.10.14 is a separate one and is discussed here: apache#13802

This issue was resolved but the fix went in the 1.10.15 version. I took the same fix: https://github.com/apache/airflow/pull/13803/files and have applied it in this commit (5e87232).

The resolution worked and I did the testing on gke devel cluster: https://airflow-smantri--devel--etl-workflow-users.service.qus1.twitter.biz/admin/airflow/tree?dag_id=bq_dal_integration_dag

Prior to the fix, if the max_active_runs = 1, and if you start more than 1 run simultaneously, all of them will start their tasks, and would not honour max_active_runs configuration.

With this fix, if the max_active_runs = 1, and one of the runs is going on while the second run is triggered, the second run starts with the start_date being the time when it is triggered, but the second run won't start any of its tasks until the prior run is complete.